+++ Userland/Libraries/LibGfx/ICC/TagTypes.cpp
@@ -6,6 +6,7 @@
 
 #include <AK/DeprecatedString.h>
 #include <AK/Endian.h>
+#include <LibGfx/ICC/BinaryFormat.h>
 #include <LibGfx/ICC/TagTypes.h>
 #include <LibGfx/ICC/Tags.h>
 #include <LibTextCodec/Decoder.h>
@@ -13,57 +14,6 @@
 namespace Gfx::ICC {
 
 namespace {
-
-// ICC V4, 4.6 s15Fixed16Number
-using s15Fixed16Number = i32;
-
-// ICC V4, 4.7 u16Fixed16Number
-using u16Fixed16Number = u32;
-
-// ICC V4, 4.14 XYZNumber
-struct XYZNumber {
-    BigEndian<s15Fixed16Number> x;
-    BigEndian<s15Fixed16Number> y;
-    BigEndian<s15Fixed16Number> z;
-
-    operator XYZ() const
-    {
-        return XYZ { x / (double)0x1'0000, y / (double)0x1'0000, z / (double)0x1'0000 };
-    }
-};
-
-// Common bits of ICC v4, Table 40 — lut16Type encoding and Table 44 — lut8Type encoding
-struct LUTHeader {
-    u8 number_of_input_channels;
-    u8 number_of_output_channels;
-    u8 number_of_clut_grid_points;
-    u8 reserved_for_padding;
-    BigEndian<s15Fixed16Number> e_parameters[9];
-};
-static_assert(AssertSize<LUTHeader, 40>());
-
-// Common bits of ICC v4, Table 45 — lutAToBType encoding and Table 47 — lutBToAType encoding
-struct AdvancedLUTHeader {
-    u8 number_of_input_channels;
-    u8 number_of_output_channels;
-    BigEndian<u16> reserved_for_padding;
-    BigEndian<u32> offset_to_b_curves;
-    BigEndian<u32> offset_to_matrix;
-    BigEndian<u32> offset_to_m_curves;
-    BigEndian<u32> offset_to_clut;
-    BigEndian<u32> offset_to_a_curves;
-};
-static_assert(AssertSize<AdvancedLUTHeader, 24>());
-
-// ICC v4, Table 46 — lutAToBType CLUT encoding
-// ICC v4, Table 48 — lutBToAType CLUT encoding
-// (They're identical.)
-struct CLUTHeader {
-    u8 number_of_grid_points_in_dimension[16];
-    u8 precision_of_data_elements; // 1 for u8 entries, 2 for u16 entries.
-    u8 reserved_for_padding[3];
-};
-static_assert(AssertSize<CLUTHeader, 20>());
 
 ErrorOr<void> check_reserved(ReadonlyBytes tag_bytes)
 {
